Remove protocol.py
Issue https://github.com/BD2KGenomics/ga4gh-integration/issues/53
