+++ ga4gh/server/converters.py
@@ -11,7 +11,8 @@ import collections
 import pysam
 
 import ga4gh.server.datamodel.reads as reads
-import ga4gh.server.protocol as protocol
+
+import ga4gh.schemas.protocol as protocol
 
 
 class AbstractConverter(object):
