+++ ga4gh/server/datamodel/__init__.py
@@ -13,7 +13,8 @@ import json
 import os
 
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
+
+import ga4gh.schemas.protocol as protocol
 
 
 class PysamFileHandleCache(object):
