+++ ga4gh/server/datamodel/bio_metadata.py
@@ -9,8 +9,9 @@ import datetime
 import json
 
 import ga4gh.server.datamodel as datamodel
-import ga4gh.server.protocol as protocol
 import ga4gh.server.exceptions as exceptions
+
+import ga4gh.schemas.protocol as protocol
 
 
 class Biosample(datamodel.DatamodelObject):
