+++ ga4gh/server/datamodel/datasets.py
@@ -10,12 +10,12 @@ import ga4gh.server.datamodel.reads as reads
 import ga4gh.server.datamodel.sequence_annotations as sequence_annotations
 import ga4gh.server.datamodel.variants as variants
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
 import ga4gh.server.datamodel.bio_metadata as biodata
 import ga4gh.server.datamodel.genotype_phenotype as g2p
 import ga4gh.server.datamodel.rna_quantification as rnaQuantification
 
 import ga4gh.schemas.pb as pb
+import ga4gh.schemas.protocol as protocol
 
 
 class Dataset(datamodel.DatamodelObject):
