+++ ga4gh/server/datamodel/genotype_phenotype_featureset.py
@@ -12,9 +12,10 @@ import rdflib
 from rdflib import RDF
 
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
 import ga4gh.server.datamodel.sequence_annotations as sequence_annotations
 import ga4gh.server.datamodel.genotype_phenotype as g2p
+
+import ga4gh.schemas.protocol as protocol
 
 # annotation keys
 TYPE = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'
