+++ ga4gh/server/datamodel/reads.py
@@ -16,9 +16,9 @@ import pysam
 import ga4gh.server.datamodel as datamodel
 import ga4gh.server.datamodel.references as references
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
 
 import ga4gh.schemas.pb as pb
+import ga4gh.schemas.protocol as protocol
 
 
 def parseMalformedBamHeader(headerDict):
