+++ ga4gh/server/datamodel/rna_quantification.py
@@ -7,9 +7,10 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import ga4gh.server.datamodel as datamodel
-import ga4gh.server.protocol as protocol
 import ga4gh.server.exceptions as exceptions
 import ga4gh.server.sqlite_backend as sqlite_backend
+
+import ga4gh.schemas.protocol as protocol
 
 
 """
