+++ scripts/server_benchmark.py
@@ -15,9 +15,10 @@ import guppy
 
 import glue
 
+import ga4gh.schemas.protocol as protocol
+
 glue.ga4ghImportGlue()
 import ga4gh.server.backend as backend  # noqa
-import ga4gh.server.protocol as protocol  # noqa
 import ga4gh.server.datarepo as datarepo  # noqa
