+++ tests/datadriven/test_sequence_annotations.py
@@ -10,9 +10,10 @@ import ga4gh.server.datamodel as datamodel
 import ga4gh.server.datamodel.datasets as datasets
 import ga4gh.server.datamodel.references as references
 import ga4gh.server.datamodel.sequence_annotations as sequence_annotations
-import ga4gh.server.protocol as protocol
 import tests.datadriven as datadriven
 import tests.paths as paths
+
+import ga4gh.schemas.protocol as protocol
 
 _datasetName = "ds"
