+++ tests/datadriven/test_variants.py
@@ -15,12 +15,12 @@ import ga4gh.server.datamodel as datamodel
 import ga4gh.server.datamodel.datasets as datasets
 import ga4gh.server.datamodel.references as references
 import ga4gh.server.datamodel.variants as variants
-import ga4gh.server.protocol as protocol
 import ga4gh.server.exceptions as exceptions
 import tests.datadriven as datadriven
 import tests.paths as paths
 
 import ga4gh.common.utils as utils
+import ga4gh.schemas.protocol as protocol
 
 
 def testVariantSets():
