+++ tests/end_to_end/test_g2p.py
@@ -4,9 +4,10 @@ G2P testing on the test data
 import unittest
 
 import ga4gh.server.datamodel as datamodel
-import ga4gh.server.protocol as protocol
 import ga4gh.server.frontend as frontend
 import tests.paths as paths
+
+import ga4gh.schemas.protocol as protocol
 
 
 class TestG2P(unittest.TestCase):
