+++ tests/end_to_end/test_repo_manager.py
@@ -14,7 +14,7 @@ import json
 import ga4gh.server.cli.repomanager as cli_repomanager
 import tests.paths as paths
 
-import ga4gh.server.protocol as protocol
+import ga4gh.schemas.protocol as protocol
 
 
 class RepoManagerEndToEndTest(unittest.TestCase):
