+++ tests/unit/test_faultydata.py
@@ -11,7 +11,8 @@ import unittest
 import ga4gh.server.datamodel.datasets as datasets
 import ga4gh.server.datamodel.variants as variants
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
+
+import ga4gh.schemas.protocol as protocol
 
 
 class FaultyVariantDataTest(unittest.TestCase):
