+++ tests/unit/test_protocol_errors.py
@@ -9,7 +9,8 @@ import unittest
 
 import ga4gh.server.frontend as frontend
 import ga4gh.server.exceptions as exceptions
-import ga4gh.server.protocol as protocol
+
+import ga4gh.schemas.protocol as protocol
 
 
 class TestFrontendErrors(unittest.TestCase):
