+++ tests/unit/test_views.py
@@ -12,7 +12,8 @@ import tests.paths as paths
 
 import ga4gh.server.datamodel as datamodel
 import ga4gh.server.frontend as frontend
-import ga4gh.server.protocol as protocol
+
+import ga4gh.schemas.protocol as protocol
 
 
 class TestFrontend(unittest.TestCase):
